# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task API Documentation",
        "description": "Test task API Documentation Description",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://task.firscoders.shop/api/",
            "description": "Global API server"
        },
        {
            "url": "http://test-task.loc/api/",
            "description": "Local API server"
        }
    ],
    "paths": {
        "/v1/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "description": "User login api.",
                "operationId": "authLogin",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "User's phone number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "+998912223344"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user12345"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!"
                    },
                    "422": {
                        "description": "Validation error!"
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                }
            }
        },
        "/v1/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "description": "User logout.",
                "operationId": "userLogout",
                "responses": {
                    "200": {
                        "description": "Success!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}